Rename TITANIC DATASET.ipynb to Predicting Titanic Survivors.ipynb
This project analyses data from the tragic Titanic event and also builds models to determine if a passenger survived. 
It involves: Understanding the data, Exploratory data analysis, Data Pre-processing, Splitting the data into train and test data and building four models in order to determine which of the models built is better suited for analysing passenger survival.